CreditCardVault/requirements.txt
customtkinter==5.2.2
mysql_connector_repackaged==0.3.1
pycryptodome==3.20.0
